+++ src/analyzers/two-fer/index.ts
@@ -1,22 +1,24 @@
-import {
-  BinaryExpression,
-  ConditionalExpression,
-  IfStatement,
-  LogicalExpression,
-  Program,
-  TemplateLiteral,
-  Parameter
-} from "@typescript-eslint/typescript-estree/dist/ts-estree/ts-estree"
-import { AST_NODE_TYPES } from "@typescript-eslint/typescript-estree"
+import { AST_NODE_TYPES } from "@typescript-eslint/typescript-estree";
+import { ConditionalExpression, IfStatement, LogicalExpression, Parameter, Program, TemplateLiteral } from "@typescript-eslint/typescript-estree/dist/ts-estree/ts-estree";
 
-import { AnalyzerImpl } from "../AnalyzerImpl"
-
-import { extractAll } from "../utils/extract_all"
-import { extractDefaultExport } from "../utils/extract_export"
-import { extractFirst } from "../utils/extract_first"
-import { extractMainMethod, MainMethod } from "../utils/extract_main_method"
-
-import { factory } from "../../comments/comment"
+import { AnalyzerImpl } from "~src/analyzers/AnalyzerImpl";
+import { extractAll } from "~src/analyzers/utils/extract_all";
+import { extractDefaultExport } from "~src/analyzers/utils/extract_export";
+import { extractFirst } from "~src/analyzers/utils/extract_first";
+import { extractMainMethod, MainMethod } from "~src/analyzers/utils/extract_main_method";
+import { parameterName } from '~src/analyzers/utils/extract_parameter';
+import { isAssignmentPattern } from "~src/analyzers/utils/is_assignment_pattern";
+import { isBinaryExpression } from "~src/analyzers/utils/is_binary_expression";
+import { isIdentifier } from "~src/analyzers/utils/is_identifier";
+import { isLiteral } from "~src/analyzers/utils/is_literal";
+import { isLogicalExpression } from "~src/analyzers/utils/is_logical_expression";
+import { isTemplateLiteral } from "~src/analyzers/utils/is_template_literal";
+import { isUnaryExpression } from "~src/analyzers/utils/is_unary_expression";
+import { parameterType } from "~src/analyzers/utils/type_annotations";
+import { factory } from "~src/comments/comment";
+import { AstParser, ParsedSource } from "~src/parsers/AstParser";
+import { NoSourceError } from "~src/errors/NoSourceError";
+import { ParserError } from "~src/errors/ParserError";
 import {
   NO_METHOD,
   NO_DEFAULT_EXPORT,
@@ -26,20 +28,19 @@ import {
   PREFER_STRICT_EQUALITY,
   NO_PARAMETER,
   PREFER_EXPLICIT_RETURN_TYPE_FOR_PUBLIC_API,
-  UNEXPECTED_BOXED_TYPE
+  UNEXPECTED_BOXED_TYPE,
+  PARSE_ERROR
 } from "../../comments/shared"
 
 import { isRequired, isOptional } from "../utils/test_parameter"
-import { parameterName } from '../utils/extract_parameter'
-import { parameterType, annotateType } from "../utils/type_annotations"
-import { isAssignmentPattern } from "../utils/is_assignment_pattern";
-import { isBinaryExpression } from "../utils/is_binary_expression";
-import { isIdentifier } from "../utils/is_identifier";
-import { isLiteral } from "../utils/is_literal";
-import { isTemplateLiteral } from "../utils/is_template_literal";
-import { isUnaryExpression } from "../utils/is_unary_expression";
-import { isLogicalExpression } from "../utils/is_logical_expression";
-import { AstParser } from "../../parsers/AstParser";
+
+/**
+ * The factories here SHOULD be kept in sync with exercism/website-copy. Under
+ * normal use, they do NOT dictate the actual commentary output of the analyzer,
+ * as that is provided by the website-copy repo.
+ *
+ * https://github.com/exercism/website-copy/tree/master/automated-comments/typescript/two-fer
+ */
 
 const OPTIMISE_DEFAULT_VALUE = factory<'parameter'>`
 You currently use a conditional to branch in case there is no value passed
@@ -47,8 +48,8 @@ into twoFer, but instead you could set the default value of ${'parameter'} to
 'you' to avoid this conditional.
 `('typescript.two-fer.optimise_default_value')
 
-const OPTIMISE_EXPLICIT_DEFAULT_VALUE = factory<'parameter' | 'maybe_undefined_expression'>`
-Instead of relying on ${'maybe_undefined_expression'} being "undefined" when
+const OPTIMISE_EXPLICIT_DEFAULT_VALUE = factory<'parameter' | 'maybe-undefined-expression'>`
+Instead of relying on ${'maybe-undefined-expression'} being "undefined" when
 no value is passed in, you could set the default value of '${'parameter'}' to
 'you'.
 `('typescript.two-fer.optimise_explicity_default_value')
@@ -56,6 +57,11 @@ no value is passed in, you could set the default value of '${'parameter'}' to
 const REDIRECT_INCORRECT_STRING_TEMPLATE = factory`
 The string template looks incorrect. Expected a template with 3 components.
 `('typescript.two-fer.redirect_incorrect_string_template')
+
+const TIP_EXPORT_INLINE = factory`
+Did you know that you can export functions, classes and constants directly
+inline?
+`('typescript.two-fer.export_inline')
 
 const Parser: AstParser = new AstParser(undefined, 1)
 
@@ -69,21 +75,21 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
 
   private _mainParameter!: Parameter
 
-  get mainMethod() {
+  private get mainMethod(): ReturnType<typeof extractMainMethod> {
     if (!this._mainMethod) {
       this._mainMethod = extractMainMethod(this.program, 'twoFer')
     }
     return this._mainMethod
   }
 
-  get mainExport() {
+  private get mainExport(): ReturnType<typeof extractDefaultExport> {
     if (!this._mainExport) {
       this._mainExport = extractDefaultExport(this.program)
     }
     return this._mainExport
   }
 
-  get mainParameter() {
+  private get mainParameter(): Parameter {
     if (!this._mainParameter) {
       this._mainParameter = this.mainMethod!.params[0]
     }
@@ -92,7 +98,7 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
   }
 
   public async execute(input: Input): Promise<void> {
-    const [parsed] = await Parser.parse(input)
+    const [parsed] = await this.parse(input)
 
     this.program = parsed.program
     this.source = parsed.source
@@ -133,14 +139,33 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     // The solution is automatically referred to the mentor if it reaches this
   }
 
-  private checkStructure() {
+  private async parse(input: Input): never | Promise<ParsedSource[]> {
+    try {
+      return await Parser.parse(input)
+    } catch (err) {
+      if (err instanceof NoSourceError) {
+        this.logger.error(`=> [NoSourceError] ${err.message}`)
+        this.redirect()
+      }
+
+      if (err instanceof ParserError) {
+        this.logger.error(`=> [ParserError] ${err.message}`)
+        const { message, ...details } = err.original
+        this.disapprove(PARSE_ERROR({ error: message, details: JSON.stringify(details) }))
+      }
+
+      throw err
+    }
+  }
+
+  private checkStructure(): void | never {
     const method = this.mainMethod
     const [declaration,] = this.mainExport
 
     // First we check that there is a two-fer function and that this function
     // is exported.
     if (!method) {
-      this.comment(NO_METHOD({ method_name: 'twoFer' }))
+      this.comment(NO_METHOD({ 'method.name': 'twoFer' }))
     }
 
     if (!declaration) {
@@ -154,12 +179,12 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     }
   }
 
-  private checkSignature() {
+  private checkSignature(): void | never {
     const method: MainMethod = this.mainMethod!
 
     // If there is no parameter then this solution won't pass the tests.
     if (method.params.length === 0) {
-      this.disapprove(NO_PARAMETER({ function_name: method.id!.name }))
+      this.disapprove(NO_PARAMETER({ 'function.name': method.id!.name }))
     }
 
     const firstParameter = this.mainParameter!
@@ -168,8 +193,8 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     // then this won't pass the tests.
     if (isRequired(firstParameter)) {
       this.disapprove(UNEXPECTED_REQUIRED_PARAMETER({
-        parameter_name: parameterName(firstParameter),
-        parameter_type: parameterType(firstParameter)
+        'parameter.name': parameterName(firstParameter),
+        'parameter.type': parameterType(firstParameter)
       }))
     }
 
@@ -177,20 +202,17 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     // use a real parameter.
     if (firstParameter.type === AST_NODE_TYPES.RestElement) {
       const splatArgName = parameterName(firstParameter)
-      const splatArgType = annotateType(firstParameter.typeAnnotation)
+      const splatArgType = parameterType(firstParameter)
 
-      this.disapprove(UNEXPECTED_SPLAT_ARGS({ 'splat_arg_name': splatArgName, parameter_type: splatArgType }))
+      this.disapprove(UNEXPECTED_SPLAT_ARGS({
+        'splat-arg.name': splatArgName,
+        'parameter.type': splatArgType
+      }))
     }
 
     // If they use a boxed type, bail with a message.
-    const typeAnnotation = (
-           firstParameter.type === AST_NODE_TYPES.Identifier
-        && firstParameter.typeAnnotation
-      ) || (
-           firstParameter.type === AST_NODE_TYPES.AssignmentPattern
-        && firstParameter.left.type === AST_NODE_TYPES.Identifier
-        && firstParameter.left.typeAnnotation
-      )
+    const typeAnnotation = (isIdentifier(firstParameter) && firstParameter.typeAnnotation)
+        || (isAssignmentPattern(firstParameter) && isIdentifier(firstParameter.left) && firstParameter.left.typeAnnotation)
 
     if (typeAnnotation && typeAnnotation.typeAnnotation.type === AST_NODE_TYPES.TSTypeReference) {
       const boxedType = typeAnnotation.typeAnnotation.typeName.type === AST_NODE_TYPES.Identifier
@@ -202,15 +224,14 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
         this.redirect()
       } else {
         this.disapprove(UNEXPECTED_BOXED_TYPE({
-          boxed_type: boxedType,
-          literal_type: boxedType[0].toLowerCase() + boxedType.slice(1)
+          'boxed.type': boxedType,
+          'literal.type': boxedType[0].toLowerCase() + boxedType.slice(1)
         }))
       }
     }
 
     // If they have not given a return type, let's suggest they add it.
-    if (
-        !method.returnType
+    if (!method.returnType
       || method.returnType.typeAnnotation.type !== AST_NODE_TYPES.TSStringKeyword
     ) {
       // TODO better signature from actual name and if assignment, add that too
@@ -220,7 +241,7 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     }
   }
 
-  private checkForOptimalSolutions() {
+  private checkForOptimalSolutions(): void | never {
     // The optional solution looks like this:
     //
     // export function twoFer(name = 'you') {
@@ -233,8 +254,7 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     //
     // NOTE: the current tests are incorrect and want you to do name || 'you'
 
-    if (
-         !this.isDefaultArgumentOptimal()
+    if (!this.isDefaultArgumentOptimal()
       || !this.isOneLineSolution()
       || !this.isUsingTemplatedString()
     ) {
@@ -249,10 +269,17 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
       this.redirect(REDIRECT_INCORRECT_STRING_TEMPLATE())
     }
 
+    this.checkForTips()
     this.approve()
   }
 
-  private checkForApprovableSolutions() {
+  private checkForTips(): void | never {
+    if (!this.hasInlineExport()) {
+      this.comment(TIP_EXPORT_INLINE())
+    }
+  }
+
+  private checkForApprovableSolutions(): void | never {
     // If we don't have a correct default argument or a one line
     // solution then let's just get out of here.
     if (!this.isOneLineSolution()) {
@@ -280,27 +307,20 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     }
   }
 
-  private checkForTips() {
-    // if (!this.hasInlineExport()) {
-    //   this.comment(TIP_EXPORT_INLINE())
-    // }
-  }
-
-  private checkForSolutionWithoutStringTemplate() {
-    const [expression] = extractAll<BinaryExpression>(this.mainMethod!, AST_NODE_TYPES.BinaryExpression)
+  private checkForSolutionWithoutStringTemplate(): void | never {
+    const [expression] = extractAll(this.mainMethod!, AST_NODE_TYPES.BinaryExpression)
 
     //
     // "One for " + name + ", one for me."
     //
     if (isBinaryExpression(expression, '+') && (
-           expression.left.type === AST_NODE_TYPES.Literal
-        || expression.right.type === AST_NODE_TYPES.Literal
-      )) {
+      expression.left.type === AST_NODE_TYPES.Literal
+      || expression.right.type === AST_NODE_TYPES.Literal)) {
       this.disapprove(PREFER_TEMPLATED_STRINGS())
     }
   }
 
-  private checkForSolutionWithFalsyDefault() {
+  private checkForSolutionWithFalsyDefault(): void | never {
     //
     // "One for " + (name || 'you') + ", one for me."
     // `One for ${name || 'you'}, one for me.`
@@ -308,8 +328,7 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     const expression = extractFirst<LogicalExpression>(this.mainMethod!, AST_NODE_TYPES.LogicalExpression)
 
     if (expression && isLogicalExpression(expression, '||') && isIdentifier(expression.left)) {
-      if (
-           isLiteral(expression.right, 'you')
+      if (isLiteral(expression.right, 'you')
         || isTemplateLiteral(expression.right, ['you'])
       ) {
         const firstParameter = this.mainMethod!.params[0]
@@ -317,7 +336,7 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
 
         this.comment(OPTIMISE_EXPLICIT_DEFAULT_VALUE({
           parameter,
-          maybe_undefined_expression: expression.left.name
+          'maybe-undefined-expression': expression.left.name
         }))
       }
 
@@ -326,35 +345,30 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
 
     // `One for ${name ? name : 'you'}, one for me.`
     const conditionalExpression = extractFirst<ConditionalExpression>(this.mainMethod!, AST_NODE_TYPES.ConditionalExpression)
-    if (
-         conditionalExpression
+    if (conditionalExpression
       && isIdentifier(conditionalExpression.test)
       && conditionalExpression.consequent.type === conditionalExpression.test.type
       && conditionalExpression.consequent.name === conditionalExpression.test.name
-     ) {
+    ) {
 
-      if (
-             isLiteral(conditionalExpression.alternate, 'you')
-          || isTemplateLiteral(conditionalExpression.alternate, ['you'])
-        )
-      {
+      if (isLiteral(conditionalExpression.alternate, 'you')
+        || isTemplateLiteral(conditionalExpression.alternate, ['you'])) {
         const firstParameter = this.mainMethod!.params[0]
         const parameter = parameterName(firstParameter, 'name')
 
         this.comment(OPTIMISE_EXPLICIT_DEFAULT_VALUE({
           parameter,
-          maybe_undefined_expression: conditionalExpression.consequent.name
+          'maybe-undefined-expression': conditionalExpression.consequent.name
         }))
       }
     }
   }
 
-  private checkForConditionalOnDefaultArgument() {
+  private checkForConditionalOnDefaultArgument(): void | never {
     const conditionalExpressions = extractAll<ConditionalExpression>(this.mainMethod!, AST_NODE_TYPES.ConditionalExpression)
     const ifStatements = extractAll<IfStatement>(this.mainMethod!, AST_NODE_TYPES.IfStatement)
 
-    if (
-         ifStatements.length === 0
+    if (ifStatements.length === 0
       && conditionalExpressions.length === 0
       || (ifStatements.length + conditionalExpressions.length > 1)
     ) {
@@ -398,6 +412,14 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
       // if (undefined == name)
       // if (null == name)
       // if ('' == name)
+      // if (name != false)
+      // if (name != undefined)
+      // if (name != null)
+      // if (name != '')
+      // if (false != name)
+      // if (undefined != name)
+      // if (null != name)
+      // if ('' != name)
       if (
         (isBinaryExpression(test, '==') || isBinaryExpression(test, '!='))
         && (isIdentifier(test.left, defaultParameterName) || isIdentifier(test.right, defaultParameterName))
@@ -418,8 +440,7 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     // !name ? 'you' : 'name'
     // name ? name : `you`
     // !name ? `you` : 'name'
-    if (
-         (isIdentifier(consequent) && (isLiteral(alternate, 'you') || isTemplateLiteral(alternate)))
+    if ((isIdentifier(consequent) && (isLiteral(alternate, 'you') || isTemplateLiteral(alternate)))
       || (isIdentifier(alternate) && (isLiteral(consequent, 'you') || isTemplateLiteral(consequent)))
     ) {
       this.disapprove(OPTIMISE_DEFAULT_VALUE({ parameter: defaultParameterName }))
@@ -427,8 +448,7 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
 
     // name ? `One for ${name}, and one for me` : 'One for you, one for me.')
     // !name ? 'One for you, one for me.' : `One for ${name}, and one for me`
-    if (
-        (isTemplateLiteral(consequent) && (isLiteral(alternate, 'One for you, one for me.') || isTemplateLiteral(alternate)))
+    if ((isTemplateLiteral(consequent) && (isLiteral(alternate, 'One for you, one for me.') || isTemplateLiteral(alternate)))
       || isTemplateLiteral(alternate) && (isLiteral(consequent, 'One for you, one for me.') || isTemplateLiteral(consequent))
 
     ) {
@@ -439,11 +459,11 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     this.redirect()
   }
 
-  private isDefaultArgumentOptimal() {
+  private isDefaultArgumentOptimal(): boolean {
     return isAssignmentPattern(this.mainParameter, 'you')
   }
 
-  private isOneLineSolution() {
+  private isOneLineSolution(): boolean {
     // Maximum body count may be 2 (3 - 1)
     //
     // -: class TwoFer {
@@ -465,17 +485,16 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     // This trick actually looks to the inner exppresion instead of the entire
     // function in order to allow for comments inside the body.
     const { loc: { start: { line: lineStart }, end: { line: lineEnd } } } =
-         body.type === AST_NODE_TYPES.BlockStatement
+      body.type === AST_NODE_TYPES.BlockStatement
       && body.body.length === 1
       && body.body[0].type === AST_NODE_TYPES.ReturnStatement
-       ? body.body[0]
-       : this.mainMethod!
+        ? body.body[0]
+        : this.mainMethod!
 
     return (lineEnd - lineStart) <= 2
   }
 
-
-  private hasInlineExport() {
+  private hasInlineExport(): boolean {
     // Additionally make sure the export is inline by checking if it doesn't
     // have any specifiers:
     //
@@ -488,14 +507,15 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     return !isIdentifier(this.mainExport[0]!.declaration)
   }
 
-  private isUsingTemplatedString() {
+  private isUsingTemplatedString(): TemplateLiteral | undefined {
     return extractFirst<TemplateLiteral>(this.mainMethod!, AST_NODE_TYPES.TemplateLiteral)
   }
 
-  private hasThreeComponentsInTemplateLiteral() {
+  private hasThreeComponentsInTemplateLiteral(): boolean {
     const template = extractFirst<TemplateLiteral>(this.mainMethod!, AST_NODE_TYPES.TemplateLiteral)
-    return template
-      && template.quasis.length + template.expressions.length === 3
+    return !!(
+      template && template.quasis.length + template.expressions.length === 3
+    )
   }
 }