+++ src/output/processor/FileOutput.ts
@@ -1,7 +1,7 @@
-import fs from 'fs'
+import { writeFile } from '~src/utils/fs'
 import path from 'path'
 
-import { getProcessLogger } from '../../utils/logger'
+import { getProcessLogger } from '~src/utils/logger'
 
 type FileOutputOptions = Pick<ExecutionOptions, 'output' | 'inputDir'>
 
@@ -10,11 +10,7 @@ export const FileOutput: OutputProcessor = async (previous: Promise<string>, opt
   const outputPath = getOutputPath(options)
   getProcessLogger().log(`=> writing output to ${outputPath}`)
 
-  return new Promise((resolve, reject) => {
-    fs.writeFile(outputPath, output, (err) => {
-      err ? reject(err) : resolve(output)
-    })
-  })
+  return writeFile(outputPath, output)
 }
 
 function getOutputPath({ output, inputDir }: FileOutputOptions): string {