+++ src/utils/logger.ts
@@ -1,11 +1,11 @@
 type StreamBuffer = string | Buffer | Uint8Array
-type LoggerInput = StreamBuffer | (() => StreamBuffer)
+export type LoggerInput = StreamBuffer | (() => StreamBuffer)
 
 /**
  * Log the buffer to the output
  * @param buffer
  */
-function log(buffer: LoggerInput) {
+function log(buffer: LoggerInput): void {
   process.stdout.write(buffer instanceof Function ? buffer() : buffer)
 }
 
@@ -13,7 +13,7 @@ function log(buffer: LoggerInput) {
  * Log the buffer to the error output
  * @param buffer
  */
-function error(buffer: LoggerInput) {
+function error(buffer: LoggerInput): void {
   process.stderr.write(buffer instanceof Function ? buffer() : buffer)
 }
 
@@ -27,30 +27,12 @@ function fatal(this: Logger, buffer: LoggerInput, status = 1): never {
   return process.exit(status)
 }
 
-function noop(_: LoggerInput) {}
-
-const LIVE_BINDING: { current: Logger | null } = { current: null }
-
-/**
- * Set the 'global' logger
- * @param logger
- * @returns the global logger
- */
-export function setProcessLogger(logger: Readonly<Logger>) {
-  return LIVE_BINDING.current = logger
-}
-
-/**
- * Get the 'global' logger
- */
-export function getProcessLogger(): Logger {
-  return LIVE_BINDING.current!
-}
+function noop(_: LoggerInput): void {}
 
 export interface Logger {
-  error: typeof error
-  fatal: typeof fatal
-  log: typeof log
+  error: typeof error;
+  fatal: typeof fatal;
+  log: typeof log;
 }
 
 export class Logger {
@@ -62,4 +44,23 @@ export class Logger {
       log: debug ? (console ? global.console.log : log) : noop,
     })
   }
+}
+
+const NOOP_LOGGER = new Logger({ debug: false, console: false })
+const LIVE_BINDING: { current: Logger | null } = { current: NOOP_LOGGER }
+
+/**
+ * Set the 'global' logger
+ * @param logger
+ * @returns the global logger
+ */
+export function setProcessLogger(logger: Readonly<Logger>): Readonly<Logger> {
+  return LIVE_BINDING.current = logger
+}
+
+/**
+ * Get the 'global' logger
+ */
+export function getProcessLogger(): Logger {
+  return LIVE_BINDING.current || NOOP_LOGGER
 }