+++ test/helpers/bootstrap.ts
@@ -1,7 +1,7 @@
-import { ExecutionOptionsImpl } from "../../src/utils/execution_options";
-import { ExerciseImpl } from "../../src/ExerciseImpl";
-import { BootstrapResult } from "../../src/utils/bootstrap";
-import { setProcessLogger, Logger } from "../../src/utils/logger";
+import { ExecutionOptionsImpl } from "~src/utils/execution_options";
+import { ExerciseImpl } from "~src/ExerciseImpl";
+import { BootstrapResult } from "~src/utils/bootstrap";
+import { setProcessLogger, Logger } from "~src/utils/logger";
 
 export function bootstrap({ exercise, ...overrides }: { exercise: string } & Partial<ExecutionOptions>): Omit<BootstrapResult, 'input'> {
   const options = new ExecutionOptionsImpl({
@@ -10,7 +10,8 @@ export function bootstrap({ exercise, ...overrides }: { exercise: string } & Par
     output: '__fake__',
     inputDir: '__fake__',
     dry: true,
-    templates: true,
+    noTemplates: false,
+    pretty: false,
     exercise,
     ...overrides
   })