+++ test/smoke.ts
@@ -1,9 +1,9 @@
-import { TwoFerAnalyzer } from '../src/analyzers/two-fer'
-import { run } from '../src/utils/runner'
-import { find } from '../src/analyzers/Autoload'
+import { TwoFerAnalyzer } from '~src/analyzers/two-fer'
+import { run } from '~src/utils/runner'
+import { find } from '~src/analyzers/Autoload'
 
-import { bootstrap } from './helpers/bootstrap'
-import { InlineInput } from './helpers/input/InlineInput'
+import { bootstrap } from '~test/helpers/bootstrap'
+import { InlineInput } from '~test/helpers/input/InlineInput'
 
 const { options, exercise } = bootstrap({ exercise: 'two-fer' })
 
@@ -14,7 +14,7 @@ describe('When running analysis', () => {
     const solutionContent = `
     export default class TwoFer {
       static twoFer(name: string = "you"): string {
-        return \`One for \${name\}, one for me.\`
+        return \`One for \${name}, one for me.\`
       }
     }
     `.trim()
@@ -23,7 +23,7 @@ describe('When running analysis', () => {
     const input = new InlineInput([solutionContent])
     const output = await run(analyzer, input, options)
 
-    expect(output.status).toBe('approve_as_optimal');
+    expect(output.status).toBe('approve');
     expect(output.comments.length).toBe(0);
   })
 
@@ -32,7 +32,7 @@ describe('When running analysis', () => {
     const solutionContent = `
     class TwoFer {
       static twoFer(name: string = "you") {
-        return \`One for \${name\}, one for me.\`
+        return \`One for \${name}, one for me.\`
       }
     }
 
@@ -43,7 +43,7 @@ describe('When running analysis', () => {
     const input = new InlineInput([solutionContent])
     const output = await run(analyzer, input, options)
 
-    expect(output.status).toBe('approve_with_comment');
+    expect(output.status).toBe('approve');
     expect(output.comments.length).toBeGreaterThanOrEqual(1);
   })
 
@@ -63,7 +63,7 @@ describe('When running analysis', () => {
     const input = new InlineInput([solutionContent])
     const output = await run(analyzer, input, options)
 
-    expect(output.status).toBe('disapprove_with_comment');
+    expect(output.status).toBe('disapprove');
     expect(output.comments.length).toBeGreaterThanOrEqual(1);
   })